[fix, update] fixed task create error, updated models, services, enums, config, controllers with new functions

+++ src/main/java/com/example/gerenciamentoTarefas/domain/service/UserService.java
@@ -1,5 +1,6 @@
 package com.example.gerenciamentoTarefas.domain.service;
 
+import com.example.gerenciamentoTarefas.domain.enums.UserRoles;
 import com.example.gerenciamentoTarefas.domain.exception.ResourceBadRequestException;
 import com.example.gerenciamentoTarefas.domain.exception.ResourceNotFoundException;
 import com.example.gerenciamentoTarefas.domain.model.User;
@@ -8,7 +9,7 @@ import com.example.gerenciamentoTarefas.dto.User.UserRequest;
 import com.example.gerenciamentoTarefas.dto.User.UserResponse;
 import com.example.gerenciamentoTarefas.mapper.UserMapper;
 import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -37,7 +38,8 @@ public class UserService implements ICRUDService<UserRequest, UserResponse> {
     @Override
     public List<UserResponse> findAll() {
         if(userRepository.findAll().isEmpty()) throw new ResourceNotFoundException("Não há usuario existente");
-        return userRepository.findAll().stream().map(userMapper::toDto).toList();
+        return userRepository.findAll().stream()
+                .filter(p -> p.getStatus().equals(UserRoles.USER)).map(userMapper::toDto).toList();
     }
 
     @Override
@@ -50,18 +52,26 @@ public class UserService implements ICRUDService<UserRequest, UserResponse> {
     public UserResponse update(Long id, UserRequest dto) {
         findById(id);
 
-        User user = userMapper.toRequest(dto);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(user.getStatus().equals(UserRoles.ADMIN)) throw new ResourceBadRequestException("O admin não pode atualizar");
+
         user.setId(id);
         user.setName(dto.getName());
         user.setEmail(dto.getEmail());
         user.setPassword(dto.getPassword());
+        user.setPassword(passwordEncoder.encode(dto.getPassword()));
 
         return userMapper.toDto(userRepository.save(user));
     }
 
     @Override
     public void delete(Long id) {
-        findById(id);
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Não foi possivel achar usuario com id " + id));
+
+        if(user.getStatus().equals(UserRoles.ADMIN)) throw new ResourceBadRequestException("O admin não pode se apagar");
+
         userRepository.deleteById(id);
     }
 }
